Added support for metrics being passed by bundle injection.
The MetricContributor and MetricManagerContributor classes now have to
define a clone method. This method is important to support distributed
profiling where we have to send metrics across karaf servers. Another
particularity is to hot deploy kar files containing metric bundles. In
order to enable this feature, before the BundleListClassLoader starts,
it will load the kars inside of a specific folder (set to
${karaf.home}/kar_deploy) so when the objects have to be serialized, the
karaf has already loaded the bundles containing the class definitions.