Improved ProfileField, now it can contain other ProfileFields.
This way we can use SubFields to produce different kinds of metrics. To
achieve this goal, the communication layer had to be modified, I
separated the ProfileStatus that are sent to the webapp and to the
containers (using the HDFS profiling). Now the webapp receives an object
of ProfileStatusDTO that no longer implements ProfileStatus and provides
the old ProfileStatus format, i.e., ProfileFields inside ProfileFields
are transformed into root ProfileFields by adding to its name the prefix
name of the ProfileField that contains it. This way we can use the same
webapp without modifications. The containers running for the HDFS
profiling now receive a ProfileStatusDistributedDTO which implements
ProfileStatus.